+++ GameHub/GameHub-Bridging-Header.h
@@ -1,6 +1,0 @@
-#ifndef GameHub_Bridging_Header_h
-#define GameHub_Bridging_Header_h
-
-#include "Native/Include/box64_bridge.h"
-
-#endif