+++ MNEmulator/MNEmulator-Bridging-Header.h
@@ -1,0 +1,6 @@
+#ifndef MNEmulator_Bridging_Header_h
+#define MNEmulator_Bridging_Header_h
+
+#include "Native/Include/box64_bridge.h"
+
+#endif